add profileContainer some code improvements

+++ MCLauncher/Model/FileManager.cs
@@ -1,7 +1,10 @@
-﻿using System.Diagnostics;
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
 using System.Net;
 using Ionic.Zip;
+using MCLauncher.Properties;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
@@ -38,18 +41,16 @@ namespace MCLauncher.Model
         public TResult ParseWebJson<TResult>(string url)
         {
             var jObject = DownloadJson(url);
-            
+
             return jObject.ToObject<TResult>();
         }
 
         public void ExtractToDirectory(string sourceArchive, string destinationDirectory)
         {
             using (var zip = ZipFile.Read(sourceArchive))
-            { 
+            {
                 foreach (var entry in zip)
-                {
                     entry.Extract(destinationDirectory, ExtractExistingFileAction.OverwriteSilently);
-                }
             }
         }
 
@@ -60,15 +61,9 @@ namespace MCLauncher.Model
 
         public void Delete(string path)
         {
-            if (FileExist(path))
-            {
-                File.Delete(path);
-            }
+            if (FileExist(path)) File.Delete(path);
 
-            if (DirectoryExist(path))
-            {
-                Directory.Delete(path, true);
-            }
+            if (DirectoryExist(path)) Directory.Delete(path, true);
         }
 
         public bool DirectoryExist(string path)
@@ -84,6 +79,69 @@ namespace MCLauncher.Model
         public void CreateDirectory(string directory)
         {
             Directory.CreateDirectory(directory);
+        }
+
+        public void SaveProfile(Profile profile)
+        {
+            if (Settings.Default.ProfileContainer == null)
+                Settings.Default.ProfileContainer = new ProfileContainer();
+
+            foreach (var profileFromContainer in Settings.Default.ProfileContainer.Profiles)
+                if (profile.Name == profileFromContainer.Name)
+                    return;
+            Settings.Default.ProfileContainer.Profiles.Add(profile);
+
+            Settings.Default.Save();
+        }
+
+        public List<Profile> GetProfiles()
+        {
+            if (Settings.Default.ProfileContainer == null)
+                return new List<Profile>();
+
+            return Settings.Default.ProfileContainer.Profiles;
+        }
+
+        public void EditProfile(string oldProfileName, Profile newProfile)
+        {
+            if (Settings.Default.ProfileContainer == null)
+                throw new Exception("cant edit profile in null container");
+
+            for (var i = 0; i < Settings.Default.ProfileContainer.Profiles.Count; i++)
+            {
+                if (Settings.Default.ProfileContainer.Profiles[i].Name != oldProfileName) continue;
+
+                Settings.Default.ProfileContainer.Profiles[i] = newProfile;
+                Settings.Default.Save();
+                return;
+            }
+        }
+
+        public Profile GetLastProfile()
+        {
+            if (Settings.Default.ProfileContainer == null || string.IsNullOrEmpty(Settings.Default.LastProfileName))
+                return null;
+
+            foreach (var profileFromContainer in Settings.Default.ProfileContainer.Profiles)
+                if (profileFromContainer.Name == Settings.Default.LastProfileName)
+                    return profileFromContainer;
+
+            return null;
+        }
+
+        public void SaveLastProfileName(string name)
+        {
+            Settings.Default.LastProfileName = name;
+
+            Settings.Default.Save();
+        }
+
+        public string GetLastProfileName()
+        {
+            if (string.IsNullOrEmpty(Settings.Default.LastProfileName))
+                return string.Empty;
+
+            return Settings.Default.LastProfileName;
         }
     }
 }
